+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### 0.2.1
 - Fix picklecoder
-- Fix connection failure transparency
+- Fix connection failure transparency and add logging
 - Add Cache-Control and ETag on first response
+- Support Async RedisCluster client from redis-py
 
 ### 0.2.0